Add FromBytes::zero_and_get_bytes
Add `FromBytes::zero_and_get_bytes`, which zeroes the receiver and
returns the receiver's bytes as a `&mut [u8]`. Zeroing is necessary
since, thanks to the lack of `AsBytes` bound, there's no guarantee that
there are not uninitialized bytes (e.g. as inter-field padding). Also
add `zero_and_get_byte_array`, which is similar, but only works for
`Sized` types, and returns a `ByteArray<Self>`.